Refactor to accomadate redux, including comments for path forward
