+++ src/Login.jsx
@@ -1,12 +1,29 @@
 import React from 'react';
+import axios from 'axios';
+import { useNavigate } from 'react-router';
+import { Inventory } from '../server/models';
+import LoginForm from './components/LoginPage/LoginForm';
 
 
 
 const Login = () => {
+  //NEW
 
+  const navigate = useNavigate();
 
+  const handleLogin = async (event, formData) => {
+    event.preventDefault();
 
+    const res = await axios.post('/api/auth/login', formData);
 
+    if (res.data.success) {
+
+      //dispatch this to isAdmin
+      navigate('/shop');
+    }
+  };
+
+  //NEW
   return (
     <div className="flex min-h-full flex-col justify-center px-6 py-12 lg:px-8">
       <div className="sm:mx-auto sm:w-full sm:max-w-sm">
@@ -15,50 +32,21 @@ const Login = () => {
         </h2>
       </div>
 
-      <div className="mt-10 sm:mx-auto sm:w-full sm:max-w-sm">
+      <LoginForm />
 
-        <form className="space-y-6" action="#" method="POST">
-          <div>
-            <label for="username" className="block text-sm font-medium leading-6 text-gray-900">Username</label>
-            <div className="mt-2">
-              <input
-                id="username"
-                name="username"
-                type="text"
-                required className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-              >
-              </input>
-            </div>
-          </div>
-
-          <div>
-            <div className="flex items-center justify-between">
-              <label for="password" className="block text-sm font-medium leading-6 text-gray-900">Password</label>
-            </div>
-            <div className="mt-2">
-              <input
-                id="password"
-                name="password"
-                type="password"
-                required className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
-              >
-              </input>
-            </div>
-          </div>
-
-          <div>
-            <button
-              type="submit"
-              className="flex w-full justify-center rounded-md bg-indigo-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
-            >
-              Sign in
-            </button>
-          </div>
-        </form>
-
-      </div>
     </div>
   );
 }
 
 export default Login;
+
+
+
+///isadmin redux use ternary to nav to admin page or shopping page
+// {
+// isadmin ?
+// nav to admin Inventory
+// :
+// nav to shopping page
+// }
+// store logged into session, at login, send user information to redux, store isadmin separate
