+++ src/main.jsx
@@ -1,35 +1,18 @@
-import { StrictMode } from 'react'
 import React from 'react'
 import ReactDOM from "react-dom/client"
-import { createRoot } from 'react-dom/client'
-import App from './App.jsx'
 import './index.css'
-import { RouterProvider, Route, createBrowserRouter, createRoutesFromElements } from 'react-router-dom'
-import Storefront from './Storefront.jsx'
-import Login from './Login.jsx'
-import ProfilePage from './Profile.jsx'
-import AdminInventory from './AdminInventory.jsx'
 import globalStore from './store/store.js'
 import { Provider } from 'react-redux'
-
-
-const router = createBrowserRouter(
-  createRoutesFromElements(
-    <Route path='/' element={<App />}>
-      <Route path='/shop' element={<Storefront />} />
-      <Route path='/login' element={<Login />} />
-      <Route path='/profile' element={<ProfilePage />} />
-      <Route path='/admin' element={<AdminInventory />} />
-    </Route>
-  )
-)
-
+import Router from './router.jsx'
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
     < Provider store={globalStore}>
-      <RouterProvider router={router}>
-      </RouterProvider>
+      <Router />
     </Provider>
   </React.StrictMode>,
 )
+
+//now that the global store is surrounding the router, we can pass redux info to it.
+//in router.jsx each route can validate if they are in admin role or not similar to 
+// loggedin in ratings.routes.js
